running/google_ml_sample/requirements.txt
# Requirements for all census samples.
tensorflow==1.9

# Requirements for keras example only for sample.sh.
Keras=2.2.4
gunicorn==19.9.0
pandas==0.23.4
sklearn==0.0
konlpy==0.5.1
google-cloud-storage==1.13.0
